backend-server/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-cors-headers==3.7.0
numpy
opencv-python
django-extensions==3.1.3
ipython
notebook
boto3==1.17.92
python-decouple==3.4
django-storages==1.11.1
scikit-learn>=0.24.1 # Pyxelate module
scikit-image>=0.17.1 # Pyxelate module
numba>=0.53.1 # Pyxelate module
django-rest-swagger
drf-yasg
flex
